ServidorDjango/requirements.txt
# Django y PostgreSQL
django==5.0.2
django-environ==0.11.2
asgiref==3.7.2
gunicorn==20.1.0
psycopg2==2.9.9
sqlparse==0.4.2
whitenoise==6.2.0

# Funcionalidades adicionales
resend==0.6.0
requests==2.31.0
django-hosts==5.2
django-resized==1.0.2
Pillow==9.4.0
django-tinymce==3.6.1
